Tree climbing stops after a CAA RRset is found (#4)
(not after 'CAA records' are found).

Specifying "a CAA record" in the singular seems to suggest that a CA might be permitted to (for example) consider an "issue" property whilst ignoring an "issuewild" property in the same RRset. ISTM that this sentence should say that the search is for a CAA RRset, not just for "a CAA record".