Remove Section 3 (The CAA RR Type)
The definitions in this section overlapped with the definitions in the
Mechanism section significantly and created potential for confusion and
conflict. Most of the necessary information was already in the Mechanism
section, so this is mostly a straight-up delete. In a few places I
incorporated information that was only in "The CAA RR Type".

I've also added a number of examples, switched to using
"ca1.example.net" and "ca2.example.org" as the examples, moved the
DNSSEC section into Security Considerations, and renamed "Certification
Authority Processing" to "Relevant Resource Record Set".